- default sub command - print help on execute
MISSING TESTS
