+++ .gitignore
@@ -1,6 +1,7 @@
 bin/
 obj/
 
-.idea/ 
+.idea/
 
 *.nupkg
+*.DotSettings.user
