+++ NFlags/Commands/TooManyDefaultCommandsException.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace NFlags.Commands
+{
+    public class TooManyDefaultCommandsException : Exception
+    {
+    }
+}
